Add computeScheduleAsync() / separate thread
The new function run the list scheduling algorithms in a web worker;
that is, in a separate thread.

The web workers API is a bit inconvenient unfortunately. Web workers are
not (normally) meant to be created from functions, but instead they are
created from an absolute path / URL of the script that the worker will
run. To simplify distribution, our UMD module contains both the master
and worker code.

Added some test machinery.